requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
Werkzeug==3.0.2
